sealing pipeline: Remove useless storage adapter code
